+++ packages/core/src/index.ts
@@ -1,4 +1,4 @@
 export { runCli } from './cli.js';
 export { defineWebExtendConfig as defineConfig, type WebExtendConfig } from './config.js';
 export { defineWebExtConfig, type WebExtConfig } from './runner.js';
-export type { ContentScriptConfig } from '@web-extend/manifest/types';
+export type { ContentScriptConfig } from '@web-extend/manifest/browser';
