+++ packages/manifest/package.json
@@ -5,6 +5,7 @@
   "exports": {
     ".": "./src/index.ts",
     "./types": "./src/types.ts",
+    "./browser": "./src/browser.ts",
     "./common": "./src/common.ts"
   },
   "type": "module",
