+++ packages/manifest/src/polyfill.ts
@@ -1,5 +1,6 @@
 import { isDevMode } from './common.js';
-import type { ManifestWebAccessibleResourcesC2ItemType, NormalizeManifestProps } from './types.js';
+import type { NormalizeManifestProps } from './types.js';
+import type { ManifestWebAccessibleResourcesC2ItemType } from './browser.js';
 
 function polyfillManifestBetweenBrowsers({ manifest, context }: NormalizeManifestProps) {
   if (!manifest || !context) return;
